ci: hotfix the round-7 regressions on main
CI failed on 2d2c033bd with five distinct issues. This commit fixes
the four that are mechanical; the fifth (clang-tidy redundant-cast
on the new k_iclk_hz / k_hz_per_mhz divide) is the same fix landed
in this commit.

1. main.c: rename `k_prcr_lock_all` / `k_prcr_unlock_clock_lpm` to
   `k_main_prcr_lock_all` / `k_main_prcr_unlock_clock_lpm`. The
   shorter names collide with the existing anonymous enum in
   libs/rx_core/inc/rx_gpio_constants.h:657 (`k_prcr_lock_all = 0x00`),
   which was pulled in transitively via hardware.h -> rx_bus_types.h.
   Cross-build failure: "redeclaration of enumerator 'k_prcr_lock_all'".

2. rx_drv8263.c: drop redundant (uint32_t) casts on
   `k_iclk_hz / k_hz_per_mhz` -- both enums are already uint32_t.
   clang-tidy readability-redundant-casting fires under -Werror.

3. ascii-check.yml: replace the stale exclude
   `star-rx72n-firmware/LICENSED-HARDWARE.txt` with
   `star-rx72n-firmware/libs/threadx/**` -- the file moved to
   libs/threadx/ in the previous repo cleanup but the workflow's
   exclude list wasn't updated.

4. proto-gen.yml + proto.yml: drop `cache: 'pip'` from the
   actions/setup-python step. The cache resolves
   `**/requirements.txt or **/pyproject.toml`, but the only such
   files in the repo live under the nanopb git submodule and
   actions/checkout doesn't pull submodules by default. The setup
   step was failing with "No file matched". Cache provided no
   value here anyway since these workflows install nanopb via
   pip directly.

5. star_bringup/launch/slam_mvp.launch.py: add the missing
   copyright + SPDX header so ament_copyright stops failing the
   star_bringup test target.

Verified: 60/60 firmware host tests still pass; ASCII clean.

+++ star-rx72n-firmware/libs/rx_drv8263/src/rx_drv8263.c
@@ -178,7 +178,7 @@ RX_STATIC_TESTABLE void internal_delay_us(uint32_t us)
 
   /* Cycles per microsecond derive from the authoritative ICLK constant in
    * rx72n_clock.h so any future clock-tree change tracks automatically. */
-  const uint32_t   cycles_per_us = (uint32_t)k_iclk_hz / (uint32_t)k_hz_per_mhz;
+  const uint32_t   cycles_per_us = k_iclk_hz / k_hz_per_mhz;
   volatile uint32_t cycles       = us * cycles_per_us;
   while (cycles > 0) {
     cycles--;
